+++ packages/nextra-theme-docs/src/components/head.tsx
@@ -1,7 +1,8 @@
 import React, { ReactElement } from 'react'
 import NextHead from 'next/head'
 import { useTheme } from 'next-themes'
-import { renderString, useMounted } from '../utils'
+import { useMounted } from 'nextra/hooks'
+import { renderString } from '../utils'
 import { useConfig } from '../contexts'
 
 export function Head(): ReactElement {
@@ -13,6 +14,9 @@ export function Head(): ReactElement {
   // `head` can be either FC or ReactNode. We have to directly call it if it's a
   // FC because hooks like Next.js' `useRouter` aren't allowed inside NextHead.
   const head = typeof config.head === 'function' ? config.head({}) : config.head
+  const hue = config.primaryHue
+  const { dark: darkHue, light: lightHue } =
+    typeof hue === 'number' ? { dark: hue, light: hue } : hue
 
   return (
     <NextHead>
@@ -46,6 +50,17 @@ export function Head(): ReactElement {
         name="viewport"
         content="width=device-width, initial-scale=1.0, viewport-fit=cover"
       />
+      <style>{`
+        :root {
+          --nextra-primary-hue: ${lightHue}deg;
+          --nextra-navbar-height: 4rem;
+          --nextra-menu-height: 3.75rem;
+        }
+        
+        .dark {
+          --nextra-primary-hue: ${darkHue}deg;
+        }
+      `}</style>
       {head}
     </NextHead>
   )