+++ packages/nextra-theme-docs/src/components/theme-switch.tsx
@@ -2,16 +2,18 @@ import React, { memo, ReactElement } from 'react'
 import { useTheme } from 'next-themes'
 import { Select } from './select'
 import { SunIcon, MoonIcon } from 'nextra/icons'
-import { useMounted } from '../utils'
+import { useMounted } from 'nextra/hooks'
+import { useConfig } from '../contexts'
 
 export function ThemeSwitch({ lite = true }): ReactElement {
   const { theme, setTheme, systemTheme } = useTheme()
   const renderedTheme = theme === 'system' ? systemTheme : theme
+  const config = useConfig()
   const mounted = useMounted()
 
   return (
     <Select
-      position={lite ? 'right' : 'left'}
+      position={config.i18n.length > 0 ? (lite ? 'right' : 'left') : undefined}
       onChange={option => {
         setTheme(option.key)
       }}