+++ packages/nextra-theme-docs/src/types.ts
@@ -4,6 +4,7 @@ import { ThemeProviderProps } from 'next-themes/dist/types'
 import { PageOpts } from 'nextra'
 import { Item } from './utils'
 import { TOCProps } from './components/toc'
+import { NavBarProps } from './components/navbar'
 
 export type RecursivePartial<T> = {
   [P in keyof T]?: T[P] extends (infer U)[]
@@ -47,6 +48,7 @@ export interface DocsThemeConfig {
   head: ReactNode | FC
   i18n: { direction?: string; locale: string; text: string }[]
   logo: ReactNode | FC
+  navbar: ReactNode | FC<NavBarProps>
   navigation: {
     next: boolean
     prev: boolean
@@ -59,6 +61,10 @@ export interface DocsThemeConfig {
     labels: string
     link: ReactNode | FC
   }
+  primaryHue: number | {
+    dark: number
+    light: number
+  },
   project: {
     icon: ReactNode | FC
     link: string