+++ packages/nextra/tsup.config.ts
@@ -12,31 +12,20 @@ export default defineConfig([
     target
   },
   {
-    name: 'nextra-utils',
-    entry: ['src/ssg.ts', 'src/locales.ts', 'src/context.ts'],
-    format: 'cjs',
-    external: ['next/server'],
-    dts: true,
-    target
-  },
-  {
-    name: 'nextra-loader',
-    entry: ['src/loader.ts'],
-    format: 'esm',
-    dts: true,
-    target
-  },
-  {
     name: 'nextra-esm',
     entry: [
+      'src/hooks/index.ts',
       'src/loader.ts',
       'src/compile.ts',
       'src/icons/index.ts',
-      'src/components/index.ts'
+      'src/components/index.ts',
+      'src/ssg.ts',
+      'src/locales.ts',
+      'src/context.ts'
     ],
     format: 'esm',
     dts: true,
-    target,
+    target
   },
   {
     entry: ['src/types.ts'],