+++ packages/nextra-theme-blog/CHANGELOG.md
@@ -1,5 +1,15 @@
 # nextra-theme-blog
 
+## 2.0.0-beta.24
+
+## 2.0.0-beta.23
+
+## 2.0.0-beta.22
+
+### Patch Changes
+
+- 8dab966: fix invisible copy button in code blocks
+
 ## 2.0.0-beta.21
 
 ## 2.0.0-beta.20