+++ packages/nextra/CHANGELOG.md
@@ -1,5 +1,19 @@
 # nextra
 
+## 2.0.0-beta.24
+
+### Patch Changes
+
+- a2bc427: compile `context.ts`, `ssg.ts` and `locales.ts` to esm
+
+## 2.0.0-beta.23
+
+## 2.0.0-beta.22
+
+### Patch Changes
+
+- 8dab966: fix invisible copy button in code blocks
+
 ## 2.0.0-beta.21
 
 ### Patch Changes